web-app/requirements.txt
Flask
Flask_Cors
keras
numpy
Pillow
Requests
tensorflow
uvicorn